backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.3
numpy==1.26.4
deep-translator==1.11.4
gunicorn==21.2.0
